s_posting/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
kafka-python==2.0.2
python-dotenv==1.0.1
